refactor(game-engine): flatten v2/ into spec/ top level (P10)
