+++ packages/game-engine/src/models/roles/spec/index.ts
@@ -11,11 +11,83 @@
  * Resolvers are SERVER-ONLY and located in src/resolvers/.
  */
 
-// Base types (standalone, no dependencies on v2/)
+// Base types (standalone, no cross-module dependencies)
 export * from './nightSteps.types';
 export * from './plan.types';
 export * from './schema.types';
 export * from './types';
 
-// Core (single source of truth for specs, schemas, nightSteps, nightPlan)
-export * from './v2';
+// Ability system types
+export type {
+  Ability,
+  AbilityEffect,
+  AbilityTiming,
+  ActionKind,
+  ActiveAbility,
+  BlockEffect,
+  CharmEffect,
+  CheckEffect,
+  ChooseIdolEffect,
+  CompoundSubStepDef,
+  ConfirmEffect,
+  ConvertEffect,
+  GroupRevealEffect,
+  HypnotizeEffect,
+  Immunity,
+  ImmunityKind,
+  LearnEffect,
+  MeetingConfig,
+  MimicEffect,
+  NightStepUi,
+  PassiveAbility,
+  PassiveEffectKind,
+  RecognitionConfig,
+  Resource,
+  ResourceKind,
+  SwapEffect,
+  TargetCount,
+  TargetRule,
+  TriggerCondition,
+  TriggeredAbility,
+  TriggeredEffectKind,
+  WriteSlotEffect,
+} from './ability.types';
+export { TargetConstraint } from './ability.types';
+
+// Role spec types
+export type { DeathCalcRole, NightStepDef, RoleDescription, RoleSpec } from './roleSpec.types';
+
+// Specs registry + helpers
+export {
+  getAllRoleIds,
+  getRoleDisplayAs,
+  getRoleEmoji,
+  getRoleSpec,
+  getRoleStructuredDescription,
+  isValidRoleId,
+  ROLE_SPECS,
+  type RoleId,
+} from './specs';
+
+// Night plan builder
+export { buildNightPlan, NIGHT_STEP_ORDER, type NightStepId } from './plan';
+
+// Schemas builder + registry + helpers
+export {
+  buildSchemas,
+  getAllSchemaIds,
+  getSchema,
+  isValidSchemaId,
+  type SchemaId,
+  SCHEMAS,
+} from './schemas';
+
+// Night steps registry + helpers
+export {
+  getAllStepIds,
+  getStepsByRole,
+  getStepsByRoleStrict,
+  getStepSpec,
+  getStepSpecStrict,
+  NIGHT_STEPS,
+} from './nightSteps';
