+++ src/models/roles/spec/__tests__/import-boundary.test.ts
@@ -158,14 +158,18 @@ describe('Import Boundary Enforcement', () => {
       expect(content).not.toMatch(/export\s+{.*RESOLVERS.*}/);
     });
 
-    it('spec/index.ts should export types and v2 barrel', () => {
+    it('spec/index.ts should export types, specs, schemas, plan, and nightSteps', () => {
       const specIndexPath = path.join(gameEngineSpecRoot, 'index.ts');
       const content = fs.readFileSync(specIndexPath, 'utf-8');
 
       expect(content).toContain('./types');
       expect(content).toContain('./schema.types');
       expect(content).toContain('./plan.types');
-      expect(content).toContain('./v2');
+      expect(content).toContain('./specs');
+      expect(content).toContain('./schemas');
+      expect(content).toContain('./plan');
+      expect(content).toContain('./nightSteps');
+      expect(content).not.toContain('./v2');
     });
   });
 });
